# docs/openapi.yaml
openapi: 3.1.0
info:
  title: gxp-idProvider API
  version: 0.3.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://idp.gxp.internal
security: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    basicAuth:
      type: http
      scheme: basic
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
...
    AuditLogRecord:
      type: object
      additionalProperties: false
      required:
        - id
        - actorUserId
        - action
        - resourceType
        - resourceId
        - payload
        - prevHash
        - entryHash
        - integrityVersion
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        actorUserId:
          type: [string, 'null']
          format: uuid
        action:
          type: string
        resourceType:
          type: string
        resourceId:
          type: [string, 'null']
        payload:
          type: object
          additionalProperties: true
        prevHash:
          type: [string, 'null']
        entryHash:
          type: [string, 'null']
        integrityVersion:
          type: integer
        createdAt:
          type: string
          format: date-time
    AuditLogListResponse:
      type: object
      additionalProperties: false
      required: [logs, nextCursor]
      properties:
        logs:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogRecord'
        nextCursor:
          type: [string, 'null']
    SecurityEventRecord:
      type: object
      additionalProperties: false
      required: [id, userId, eventType, payload, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: [string, 'null']
          format: uuid
        eventType:
          type: string
        payload:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    SecurityEventListResponse:
      type: object
      additionalProperties: false
      required: [events, nextCursor]
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/SecurityEventRecord'
        nextCursor:
          type: [string, 'null']
    AuditIntegrityResponse:
      type: object
      additionalProperties: false
      required: [ok, checked, firstId, lastId, brokenAt, reason]
      properties:
        ok:
          type: boolean
        checked:
          type: integer
        firstId:
          type: [string, 'null']
          format: uuid
        lastId:
          type: [string, 'null']
          format: uuid
        brokenAt:
          type: [string, 'null']
          format: uuid
        reason:
          type: [string, 'null']
    AuditExportRequest:
      type: object
      additionalProperties: false
      required: [kind, format]
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        kind:
          type: string
          enum: [audit_logs, security_events, both]
        format:
          type: string
          enum: [jsonl]
    AuditExportManifest:
      type: object
      additionalProperties: false
      required: [recordCount, sha256, generatedAt]
      properties:
        recordCount:
          type: integer
        sha256:
          type: string
        generatedAt:
          type: string
          format: date-time
    AuditExportResponse:
      type: object
      additionalProperties: false
      required: [exportId, status, manifest, format, data, truncated]
      properties:
        exportId:
          type: string
          format: uuid
        status:
          type: string
          enum: [completed]
        manifest:
          $ref: '#/components/schemas/AuditExportManifest'
        format:
          type: string
          enum: [jsonl]
        data:
          type: string
        truncated:
          type: boolean
...
paths:
...
  /v1/admin/audit/logs:
    get:
      tags: [auth-protected]
      operationId: getV1AdminAuditLogs
      summary: Search audit logs
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: actorUserId
          schema:
            type: string
            format: uuid
        - in: query
          name: action
          schema:
            type: string
        - in: query
          name: resourceType
          schema:
            type: string
        - in: query
          name: resourceId
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        '200':
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/admin/audit/security-events:
    get:
      tags: [auth-protected]
      operationId: getV1AdminAuditSecurityEvents
      summary: Search security events
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: userId
          schema:
            type: string
            format: uuid
        - in: query
          name: eventType
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - in: query
          name: cursor
          schema:
            type: string
      responses:
        '200':
          description: Security events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityEventListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/admin/audit/integrity:
    get:
      tags: [auth-protected]
      operationId: getV1AdminAuditIntegrity
      summary: Verify audit log hash chain integrity in range
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Integrity verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditIntegrityResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/admin/audit/exports:
    post:
      tags: [auth-protected]
      operationId: postV1AdminAuditExports
      summary: Export audit/security records as JSONL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditExportRequest'
      responses:
        '200':
          description: Export generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditExportResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...